requirements.txt
selenium
behave
pyyaml
webdriver-manager
allure-behave
requests
